feat: Update Google Drive integration with authentication and connection check
This commit updates the Google Drive integration by adding authentication and a connection check. It introduces two new API routes: "check-connection" and "store-token". The "check-connection" route verifies if the user is connected to Google Drive by checking for the presence of Google Drive tokens in the user's private metadata. If the user is not connected, it returns a 400 status code. The "store-token" route allows the user to store their Google Drive access token in their private metadata. It expects a POST request with the access token in the request body. If the token is successfully stored, it returns a 200 status code. Otherwise, it returns a 500 status code.

The changes also include updates to the package.json file to add the "googleapis" dependency and update other dependencies to their latest versions.