feat: Update SwapPage to pass selectedChain from WalletContext to SwapWidget
This commit modifies the SwapPage component to pass the selectedChain value from the WalletContext to the SwapWidget component. By doing this, the SwapWidget can now access the selectedChain value directly from the WalletContext, ensuring that it always has the most up-to-date information. This change improves the consistency and reliability of the SwapPage component.